# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Keeparr API",
    "version": "0.3.11",
    "description": "Keeparr tags what to keep and reports what can be deleted — it never deletes anything itself.\n\n**Authentication**: most endpoints use the `keeparr_session` cookie (set by logging in through the web UI). Endpoints marked with `X-Api-Key` security also accept the admin API key (Settings → General) in the `X-Api-Key` header — intended for automation (`/api/stats`, `/api/admin/jobs`).\n\nAll ids (`ratingKey`) are the media server's item ids (Plex ratingKey or Jellyfin/Emby item id)."
  },
  "servers": [{ "url": "/", "description": "This Keeparr instance" }],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "keeparr_session",
        "description": "Signed session cookie set by the web login."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Admin API key from Settings → General."
      }
    },
...
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/api/admin/problems": {
      "get": {
        "tags": ["admin"],
        "summary": "Paged list for one problem category",
        "parameters": [
          { "name": "type", "in": "query", "required": true, "schema": { "type": "string", "enum": ["sizeMismatch", "notInArr", "missingFromPlex", "identityMismatch", "duplicates", "arrConflicts", "zeroSize", "removedButKept", "missingIds", "diskOrphans"] } },
          { "name": "offset", "in": "query", "schema": { "type": "integer" } },
          { "name": "includeMissingIds", "in": "query", "schema": { "type": "string", "enum": ["1"] }, "description": "`notInArr` only: also list titles with no external id (hidden by default — they can never match *arr and have their own missingIds category)" },
          { "name": "sort", "in": "query", "schema": { "type": "string" }, "description": "Per-category allow-list (e.g. `title`, `size`, `added`, `delta`, `arrSize`, `instance`, `name`); unknown/absent = the category's default order" },
          { "name": "dir", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "sections", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated library ids; filters categories whose rows are media items" },
          { "name": "kind", "in": "query", "schema": { "type": "string", "enum": ["movie", "show"] } }
        ],
        "responses": {
          "200": { "description": "`{type, items[], hasMore, nextOffset}` — item shape varies per category (`duplicates` items are groups with members; `diskOrphans` items are filesystem entries `{name, sectionId, path, isDir, sizeBytes, sizeSkipped}`). Media-item rows carry `dirPath` — the full server-side folder path (null until a library scan captures it); `missingFromPlex` rows carry the *arr-side `path`." },
          "400": { "description": "`unknown_type` (missing/unknown type — no default view), `arr_not_configured` (arr-gated type without Sonarr/Radarr), or `storage_not_configured` (`diskOrphans` without storage mappings)" }
        }
      }
    }
  }
}